add depend on libjpeg62-turbo
previously on wheezy, we build qemu with libjpeg8. (qemu depend on libsdl, which depend on libjpeg)
But this package is missing is jessie (because of a last minute bug).

But that mean that upgrading from wheezy to jessie, keep the old wheezy libjpeg8. (and libjpeg8-dev).

So, it's better to add explicit depencines on libjpeg-dev and libjpeg62-turbo.  (like the official qemu debian package)
to force the upgrade of the libs.

Signed-off-by: Alexandre Derumier <aderumier@odiso.com>