Fix #932: passing BDRV_O_PROTOCOL breaks qcow2 on gluster
Passing BDRV_O_PROTOCOL causes qemu to open protocol based
paths as raw. This fails for our storage when using
glusterfs because we can use qcow2 files on there.
This also causes vma to refuse to write to them as the
expected size (the one the qcow2 was created for) does not
match the recognized file size (the size of the file
itself). (Which is good because it means the files should
not have been accessed via the wrong format backend by
accident.)

The reason for passing it was to deal with format probing
when using RBD without KRBD.

As described in the patch commit: we now provide a way to
specify the format explicitly and thereby follow qemu with
deprecating automatic guessing of raw formats.

This re-enables the raw-probing warning for non-krbd ceph
storages which now has to be addressed by passing the format
to the map fifo in PVE::QemuServer::restore_vma_archive().